How to use refactoring and also CommonLeftHeader refactor
