+++ app/components/CommonComponents/CommonLeftHeader.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import SmallHeaderBlock from 'components/SmallHeaderBlock';
+
+export default styled(SmallHeaderBlock)`
+  padding: 5px 2px;
+  border-bottom: 2px solid black;
+  margin: 0;
+`;
