+++ app/components/CommonComponents/index.js
@@ -74,13 +74,6 @@ export const BorderedButton = styled.button`
   cursor: pointer;
 `;
 
-
-export const CommonLeftHeader = styled(SmallHeaderBlock)`
-  padding: 5px 2px;
-  border-bottom: 2px solid black;
-  margin: 0;
-`;
-
 CommonComponents.propTypes = {
 
 };
