+++ app/components/ToolHowToUse/Content.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+export default  styled.div`
+  img { width: 100%; }
+  li { margin-bottom: 20px; }
+`;
