+++ app/components/ToolHowToUse/Header.js
@@ -1,0 +1,7 @@
+import CommonLeftHeader from 'components/CommonComponents/CommonLeftHeader';
+import styled from 'styled-components';
+
+export default styled(CommonLeftHeader)`
+  border: none;
+  margin-top: 40px;
+`;
