+++ app/components/ToolWhyItWorked/index.js
@@ -8,7 +8,7 @@ import React from 'react';
 import styled from 'styled-components';
 
 import { FormattedMessage, injectIntl } from 'react-intl';
-import { CommonLeftHeader } from 'components/CommonComponents';
+import CommonLeftHeader from 'components/CommonComponents/CommonLeftHeader';
 import ContentBlock from 'components/ContentBlock';
 import LanguageThemeProvider from 'components/LanguageThemeProvider';
 
@@ -36,7 +36,7 @@ function ToolWhyItWorked(props) {
   if (props.text === undefined || props.text.length == 0) return null;
 
   if (!props.show) return null;
-  
+
   return (
     <LanguageThemeProvider>
       <Header lang={props.intl.locale}>
