--- app/components/TypeMenu/messages.js
@@ -1,17 +1,0 @@
-/*
- * TypeMenu Messages
- *
- * This contains all the text for the TypeMenu component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.TypeMenu.header',
-    defaultMessage: 'This is the TypeMenu component !',
-  },
-  stories: {
-    id: 'app.components.TypeMenu.stories',
-    defaultMessage: 'Stories',
-  },
-});
