--- app/components/TypeMenu/tests/index.test.js
@@ -1,10 +1,0 @@
-// import React from 'react';
-// import { shallow } from 'enzyme';
-
-// import TypeMenu from '../index';
-
-describe('<TypeMenu />', () => {
-  it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(false);
-  });
-});
