+++ app/containers/ToolHowToUse/staticText.js
@@ -1,0 +1,6 @@
+export default {
+  header: {
+    id: 'module.how-to-use',
+    defaultMessage: 'How to Use',
+  },
+}
